+++ react-ui/src/components/Home/EsportEvents/EsportEvents.js
@@ -1,7 +1,5 @@
 import React, { Component } from 'react'
 
-
-import EsportEventsNav from './EsportEventsNav.js'
 import EventsView from './EventsView'
 
 
@@ -11,8 +9,6 @@ class EsportEvents extends Component{
 			<div id="EsportEvents">
 
 				<h2>Esport Events</h2>
-
-				<EsportEventsNav />
 
 				<EventsView />
