--- react-ui/src/components/Home/TestComponent.js
@@ -1,20 +1,0 @@
-import React, { Component } from 'react'
-
-class TestComponent extends Component{
-
-	constructor(props){
-		super(props)
-
-		this.state = {
-			isLoggedIn: false
-		}
-	}
-
-	render(){
-		return(
-			<div></div>
-		)
-	}
-}
-
-export default TestComponent
